# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImageInput {
  /** Base64 encoded image data */
  image: string;
  mediaType?: string;
}

export interface PlantIdentification {
  /** Saved scan ID (MongoDB ObjectId) */
  id?: string;
  commonName: string;
  scientificName: string;
  family: string;
  confidence: number;
  toxicityToHumans: string;
  toxicityToPets: string;
  weedPotential: string;
  aliases: string[];
  careSummary: string;
  wateringIntervalDays: number;
  fertilizingIntervalDays: number;
  sunlight?: string;
  humidity?: string;
  temperature?: string;
  soilType?: string;
}

export type DiagnosisCauseLikelihood = typeof DiagnosisCauseLikelihood[keyof typeof DiagnosisCauseLikelihood];


export const DiagnosisCauseLikelihood = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface DiagnosisCause {
  name: string;
  likelihood: DiagnosisCauseLikelihood;
  description: string;
}

export type PlantDiagnosisSeverity = typeof PlantDiagnosisSeverity[keyof typeof PlantDiagnosisSeverity];


export const PlantDiagnosisSeverity = {
  healthy: 'healthy',
  mild: 'mild',
  moderate: 'moderate',
  severe: 'severe',
} as const;

export interface PlantDiagnosis {
  isHealthy: boolean;
  causes: DiagnosisCause[];
  severity: PlantDiagnosisSeverity;
  treatmentSteps: string[];
  summary: string;
}

export interface ScanRecord {
  id: string;
  commonName: string;
  scientificName: string;
  family: string;
  confidence: number;
  aliases: string[];
  careSummary: string;
  wateringIntervalDays: number;
  fertilizingIntervalDays: number;
  /** @nullable */
  sunlight?: string | null;
  /** @nullable */
  humidity?: string | null;
  /** @nullable */
  temperature?: string | null;
  /** @nullable */
  soilType?: string | null;
  toxicityToHumans: string;
  toxicityToPets: string;
  weedPotential: string;
  scannedAt: string;
}

export interface GardenPlantInput {
  commonName: string;
  scientificName: string;
  family: string;
  aliases?: string[];
  careSummary?: string;
  wateringIntervalDays?: number;
  fertilizingIntervalDays?: number;
  /** @nullable */
  sunlight?: string | null;
  /** @nullable */
  humidity?: string | null;
  /** @nullable */
  temperature?: string | null;
  /** @nullable */
  soilType?: string | null;
  toxicityToHumans?: string;
  toxicityToPets?: string;
  /** @nullable */
  notes?: string | null;
}

export interface GardenPlant {
  id: string;
  commonName: string;
  scientificName: string;
  family: string;
  aliases?: string[];
  /** @nullable */
  careSummary?: string | null;
  /** @nullable */
  wateringIntervalDays?: number | null;
  /** @nullable */
  fertilizingIntervalDays?: number | null;
  /** @nullable */
  sunlight?: string | null;
  /** @nullable */
  humidity?: string | null;
  /** @nullable */
  temperature?: string | null;
  /** @nullable */
  soilType?: string | null;
  /** @nullable */
  toxicityToHumans?: string | null;
  /** @nullable */
  toxicityToPets?: string | null;
  /** @nullable */
  notes?: string | null;
  savedAt: string;
}

export type AppStatsTopFamiliesItem = {
  family: string;
  count: number;
};

export interface AppStats {
  totalScans: number;
  gardenSize: number;
  topFamilies: AppStatsTopFamiliesItem[];
  recentScans: ScanRecord[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  plantContext?: string;
}

export interface ChatResponse {
  content: string;
}

export interface ApiError {
  error: string;
  message: string;
}

export type ListScansParams = {
limit?: number;
};