# lib/api-zod/src/generated/types/appStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { AppStatsTopFamiliesItem } from './appStatsTopFamiliesItem';
import type { ScanRecord } from './scanRecord';

export interface AppStats {
  totalScans: number;
  gardenSize: number;
  topFamilies: AppStatsTopFamiliesItem[];
  recentScans: ScanRecord[];
}