# lib/api-zod/src/generated/types/diagnosisCause.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisCauseLikelihood } from './diagnosisCauseLikelihood';

export interface DiagnosisCause {
  name: string;
  likelihood: DiagnosisCauseLikelihood;
  description: string;
}