# lib/api-zod/src/generated/types/gardenPlant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */

export interface GardenPlant {
  id: string;
  commonName: string;
  scientificName: string;
  family: string;
  aliases?: string[];
  /** @nullable */
  careSummary?: string | null;
  /** @nullable */
  wateringIntervalDays?: number | null;
  /** @nullable */
  fertilizingIntervalDays?: number | null;
  /** @nullable */
  sunlight?: string | null;
  /** @nullable */
  humidity?: string | null;
  /** @nullable */
  temperature?: string | null;
  /** @nullable */
  soilType?: string | null;
  /** @nullable */
  toxicityToHumans?: string | null;
  /** @nullable */
  toxicityToPets?: string | null;
  /** @nullable */
  notes?: string | null;
  savedAt: Date;
}