# lib/api-zod/src/generated/types/gardenPlantInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */

export interface GardenPlantInput {
  commonName: string;
  scientificName: string;
  family: string;
  aliases?: string[];
  careSummary?: string;
  wateringIntervalDays?: number;
  fertilizingIntervalDays?: number;
  /** @nullable */
  sunlight?: string | null;
  /** @nullable */
  humidity?: string | null;
  /** @nullable */
  temperature?: string | null;
  /** @nullable */
  soilType?: string | null;
  toxicityToHumans?: string;
  toxicityToPets?: string;
  /** @nullable */
  notes?: string | null;
}