# lib/api-zod/src/generated/types/plantDiagnosis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeafLens Plant Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisCause } from './diagnosisCause';
import type { PlantDiagnosisSeverity } from './plantDiagnosisSeverity';

export interface PlantDiagnosis {
  isHealthy: boolean;
  causes: DiagnosisCause[];
  severity: PlantDiagnosisSeverity;
  treatmentSteps: string[];
  summary: string;
}